site.pp: add initial puppet configuration for berlin.freifunk.net
Creates a default nginx configuration with a vhost for berlin.freifunk.net and a
permant https redirect.

The git source is located in /usr/src/berlin.freifunk.net.

We install python and a virtual env in /usr/src/berlin.freifunk.net.

We create a cronjob that deploys the site every minute to
/srv/www/berlin.freifunk.net.